Add Perturbation schema when reading parquet files
This assists parquet file reading so that it is more independent of our current Perturbation format.
Without adding the schema, the parquet needs to be 100% similar to Perturbations.
At the moment of writing, this was needed because the parquet files only have 4 treatment types but the
Perturbation class have the full 14 types.